# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Construction Coordination Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Projects",
        "operationId": "projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array",
                  "title": "Response Projects Api Projects Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "project-lifecycle"
        ],
        "summary": "Create Project",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "project-lifecycle"
        ],
        "summary": "Open Project",
        "operationId": "open_project_api_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/sources/{source_id}/revisions/{revision_id}/import": {
      "post": {
        "tags": [
          "project-sources"
        ],
        "summary": "Import Revision",
        "operationId": "import_revision_api_projects__project_id__sources__source_id__revisions__revision_id__import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "project-sources"
        ],
        "summary": "Import Status",
        "operationId": "import_status_api_projects__project_id__sources__source_id__revisions__revision_id__import_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AgentRun"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Import Status Api Projects  Project Id  Sources  Source Id  Revisions  Revision Id  Import Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/sources": {
      "post": {
        "tags": [
          "project-sources"
        ],
        "summary": "Create Source",
        "operationId": "create_source_api_projects__project_id__sources_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectSource"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSource"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "project-sources"
        ],
        "summary": "Sources",
        "operationId": "sources_api_projects__project_id__sources_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSourceStatus"
                  },
                  "title": "Response Sources Api Projects  Project Id  Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/sources/{source_id}": {
      "get": {
        "tags": [
          "project-sources"
        ],
        "summary": "Source",
        "operationId": "source_api_projects__project_id__sources__source_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSourceStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/sources/{source_id}/revisions": {
      "get": {
        "tags": [
          "project-sources"
        ],
        "summary": "Revisions",
        "operationId": "revisions_api_projects__project_id__sources__source_id__revisions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSourceRevision"
                  },
                  "title": "Response Revisions Api Projects  Project Id  Sources  Source Id  Revisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "project-sources"
        ],
        "summary": "Upload Revision",
        "operationId": "upload_revision_api_projects__project_id__sources__source_id__revisions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_revision_api_projects__project_id__sources__source_id__revisions_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionUploadResult"
                }
              }
            }
          },
          "200": {
            "description": "Identical content already stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionUploadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/sources/{source_id}/revisions/{revision_id}": {
      "get": {
        "tags": [
          "project-sources"
        ],
        "summary": "Revision",
        "operationId": "revision_api_projects__project_id__sources__source_id__revisions__revision_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSourceRevision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/sources/{source_id}/revisions/{revision_id}/content": {
      "get": {
        "tags": [
          "project-sources"
        ],
        "summary": "Revision Content",
        "operationId": "revision_content_api_projects__project_id__sources__source_id__revisions__revision_id__content_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stored original file bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentRun": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "WAITING_APPROVAL",
              "COMPLETED",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ],
            "title": "Status",
            "default": "QUEUED"
          },
          "runtime": {
            "type": "string",
            "title": "Runtime",
            "default": "dbos"
          },
          "runtime_execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Execution Id"
          },
          "generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation",
            "default": 0
          },
          "runtime_generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Runtime Generation",
            "default": 0
          },
          "category": {
            "type": "string",
            "enum": [
              "coordination",
              "investigation",
              "engineering_recheck",
              "document_parse",
              "bim_import",
              "optimization",
              "vision",
              "embedding_index"
            ],
            "title": "Category",
            "default": "coordination"
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "event_id",
          "status",
          "runtime",
          "runtime_execution_id",
          "generation",
          "runtime_generation",
          "category",
          "analysis_id",
          "error",
          "created_at",
          "updated_at"
        ],
        "title": "AgentRun"
      },
...
      "Body_upload_revision_api_projects__project_id__sources__source_id__revisions_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "external_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 180
              },
              {
                "type": "null"
              }
            ],
            "title": "External Label"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_revision_api_projects__project_id__sources__source_id__revisions_post"
      },
...
      "CreateProject": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateProject"
      },
      "CreateProjectSource": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "BIM",
              "DOCUMENT",
              "DRAWING",
              "SCHEDULE"
            ],
            "title": "Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "kind"
        ],
        "title": "CreateProjectSource"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "timezone"
        ],
        "title": "Project"
      },
...
      "ProjectSource": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "BIM",
              "DOCUMENT",
              "DRAWING",
              "SCHEDULE"
            ],
            "title": "Kind"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "kind",
          "created_at"
        ],
        "title": "ProjectSource"
      },
      "ProjectSourceRevision": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence"
          },
          "external_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 180
              },
              {
                "type": "null"
              }
            ],
            "title": "External Label"
          },
          "original_filename": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "pattern": "^[^/\\\\\\x00-\\x1f]+$",
            "title": "Original Filename"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 180
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "size_bytes": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "import_status": {
            "type": "string",
            "const": "STORED",
            "title": "Import Status",
            "default": "STORED"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "source_id",
          "sequence",
          "external_label",
          "original_filename",
          "sha256",
          "media_type",
          "size_bytes",
          "storage_key",
          "import_status",
          "imported_at"
        ],
        "title": "ProjectSourceRevision"
      },
      "ProjectSourceStatus": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/ProjectSource"
          },
          "latest_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Revision Id"
          },
          "accepted_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted Revision Id"
          },
          "baseline_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Id"
          },
          "has_pending_revision": {
            "type": "boolean",
            "title": "Has Pending Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "latest_revision_id",
          "accepted_revision_id",
          "baseline_id",
          "has_pending_revision"
        ],
        "title": "ProjectSourceStatus"
      },
      "ProjectState": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "default": 1
          },
          "areas": {
            "items": {
              "$ref": "#/components/schemas/Area"
            },
            "type": "array",
            "title": "Areas"
          },
          "work_packages": {
            "items": {
              "$ref": "#/components/schemas/WorkPackage"
            },
            "type": "array",
            "title": "Work Packages"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRevision"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project",
          "version",
          "areas",
          "work_packages",
          "sources"
        ],
        "title": "ProjectState"
      },
...
      "RevisionUploadResult": {
        "properties": {
          "revision": {
            "$ref": "#/components/schemas/ProjectSourceRevision"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "duplicate"
        ],
        "title": "RevisionUploadResult"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}